requirements.txt
sentence-transformers>=2.2.2
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
qdrant-client== 1.11.3
pymorphy2>=0.9.1
nltk>=3.8.1
anthropic>=0.18.0
requests>=2.31.0
aiohttp>=3.9.0
wikipedia>=1.4.0
Pillow>=10.0.0
cloudinary>=1.36.0
upstash-redis>=0.15.0
redis>=5.0.0
google-cloud-translate>=3.11.0
langsmith>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dateutil>=2.8.2
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
rank_bm25>=0.2.2
langdetect==1.0.9
langchain
langchain-anthropic
pymorphy2-dicts-ru
groq>=0.33.0
prometheus-client>=0.19.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
tqdm>=4.66.1
greenlet>=3.1.1
